refactor: use primitive Button and Text
