+++ app/(tabs)/stores.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
-import { Card, Text, Button } from '@/ui/primitives';
+import { Card } from '@/ui/primitives';
+import Text from '@/ui/primitives/Text';
+import Button from '@/ui/primitives/Button';
 
 import ErrorBoundary from '@/shared/ErrorBoundary';
