+++ components/DisputeResolver.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
-import { View, Button, ActivityIndicator } from 'react-native';
+import { View, ActivityIndicator } from 'react-native';
+import Button from '@/ui/primitives/Button';
 import OrderService from '../services/orders';
 
 interface Props {
