+++ components/admin/AdminList.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { View, Text } from 'react-native';
+import { View } from 'react-native';
+import Text from '@/ui/primitives/Text';
 import { useTheme } from '../../contexts/ThemeContext';
 import Button from '@/ui/primitives/Button';
