+++ src/ui/primitives/index.ts
@@ -1,5 +1,3 @@
-export { default as Button } from './Button';
-export { default as Text } from './Text';
 export { default as Heading } from './Heading';
 export { default as TextField } from './TextField';
 export { default as Card } from './Card';
